add more linting
