+++ src/sql/platform/accounts/browser/accountPickerService.ts
@@ -7,8 +7,8 @@ import { IInstantiationService } from 'vs/platform/instantiation/common/instanti
 import { Event, Emitter } from 'vs/base/common/event';
 import * as azdata from 'azdata';
 
-import { IAccountPickerService } from 'sql/platform/accounts/common/accountPicker';
-import { AccountPicker } from 'sql/platform/accounts/browser/accountPicker';
+import { IAccountPickerService } from 'sql/platform/accounts/browser/accountPicker';
+import { AccountPicker } from 'sql/platform/accounts/browser/accountPickerImpl';
 
 export class AccountPickerService implements IAccountPickerService {
 	_serviceBrand: any;
