+++ src/sql/workbench/parts/connection/common/connectionActions.ts
@@ -13,8 +13,8 @@ import Severity from 'vs/base/common/severity';
 import { IDialogService, IConfirmation, IConfirmationResult } from 'vs/platform/dialogs/common/dialogs';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { QueryInput } from 'sql/workbench/parts/query/common/queryInput';
-import { EditDataInput } from 'sql/workbench/parts/editData/common/editDataInput';
-import { DashboardInput } from 'sql/workbench/parts/dashboard/common/dashboardInput';
+import { EditDataInput } from 'sql/workbench/parts/editData/browser/editDataInput';
+import { DashboardInput } from 'sql/workbench/parts/dashboard/browser/dashboardInput';
 import { IClipboardService } from 'vs/platform/clipboard/common/clipboardService';
 import { IQuickInputService } from 'vs/platform/quickinput/common/quickInput';
 import { IObjectExplorerService } from 'sql/workbench/services/objectExplorer/common/objectExplorerService';
