+++ src/sql/workbench/parts/dashboard/browser/dashboardEditor.ts
@@ -11,7 +11,7 @@ import { IWorkbenchThemeService } from 'vs/workbench/services/themes/common/work
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 
-import { DashboardInput } from '../common/dashboardInput';
+import { DashboardInput } from './dashboardInput';
 import { DashboardModule } from './dashboard.module';
 import { bootstrapAngular } from 'sql/platform/bootstrap/browser/bootstrapService';
 import { IDashboardComponentParams } from 'sql/platform/bootstrap/common/bootstrapParams';
