+++ src/sql/workbench/parts/dashboard/browser/dashboardInput.ts
