+++ src/sql/workbench/parts/notebook/browser/models/mimemodel.ts
@@ -3,7 +3,7 @@
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
 import { IRenderMime } from './renderMimeInterfaces';
-import { ReadonlyJSONObject } from './jsonext';
+import { ReadonlyJSONObject } from '../../common/models/jsonext';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 
 /**
