+++ src/sql/workbench/parts/notebook/browser/models/outputProcessor.ts
@@ -4,9 +4,9 @@
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
 
-import { JSONObject, isPrimitive } from './jsonext';
+import { JSONObject, isPrimitive } from '../../common/models/jsonext';
 import { MimeModel } from './mimemodel';
-import { nbformat } from './nbformat';
+import { nbformat } from '../../common/models/nbformat';
 import { nb } from 'azdata';
 
 /**
