+++ src/sql/workbench/parts/notebook/browser/outputs/plotlyOutput.component.ts
@@ -11,7 +11,7 @@ import * as types from 'vs/base/common/types';
 import { AngularDisposable } from 'sql/base/browser/lifecycle';
 import { IMimeComponent } from 'sql/workbench/parts/notebook/browser/outputs/mimeRegistry';
 import { ICellModel } from 'sql/workbench/parts/notebook/common/models/modelInterfaces';
-import { MimeModel } from 'sql/workbench/parts/notebook/common/models/mimemodel';
+import { MimeModel } from 'sql/workbench/parts/notebook/browser/models/mimemodel';
 import { getErrorMessage } from 'vs/base/common/errors';
 
 type ObjectType = object;
