+++ src/sql/workbench/parts/notebook/browser/outputs/registry.ts
@@ -3,8 +3,8 @@
 | Copyright (c) Jupyter Development Team.
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
-import { IRenderMime } from '../../common/models/renderMimeInterfaces';
-import { MimeModel } from '../../common/models/mimemodel';
+import { IRenderMime } from '../models/renderMimeInterfaces';
+import { MimeModel } from '../models/mimemodel';
 import { ReadonlyJSONObject } from '../../common/models/jsonext';
 import { defaultSanitizer } from './sanitizer';
