+++ src/sql/workbench/parts/notebook/browser/outputs/renderers.ts
@@ -4,7 +4,7 @@
 |----------------------------------------------------------------------------*/
 
 import { default as AnsiUp } from 'ansi_up';
-import { IRenderMime } from '../../common/models/renderMimeInterfaces';
+import { IRenderMime } from '../models/renderMimeInterfaces';
 import { URLExt } from '../../common/models/url';
 import { URI } from 'vs/base/common/uri';
