+++ src/sql/workbench/parts/notebook/electron-browser/outputs/markdownOutput.component.ts
@@ -16,7 +16,7 @@ import { INotebookService } from 'sql/workbench/services/notebook/common/noteboo
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { NotebookMarkdownRenderer } from 'sql/workbench/parts/notebook/electron-browser/outputs/notebookMarkdown';
-import { MimeModel } from 'sql/workbench/parts/notebook/common/models/mimemodel';
+import { MimeModel } from 'sql/workbench/parts/notebook/browser/models/mimemodel';
 import { ICellModel } from 'sql/workbench/parts/notebook/common/models/modelInterfaces';
 import { useInProcMarkdown, convertVscodeResourceToFileInSubDirectories } from 'sql/workbench/parts/notebook/common/models/notebookUtils';
 import { URI } from 'vs/base/common/uri';
